+++ docs/source/tutorial/uncertainty.rst
@@ -1,7 +1,6 @@
 Analysis of Uncertainty
 =======================
-.. todo::
-
-    @mberr please give a bit of narrative on why you would want to use the uncertainty predictions,
-    how to interpret the results (e.g., what's good/bad, what actions could be taken to make improvements),
-    relevantion citations, etc.
+.. automodapi:: pykeen.models.uncertainty
+    :no-heading:
+    :headings: --
+    :no-inheritance-diagram:
