+++ src/pykeen/models/predict.py
@@ -5,7 +5,7 @@
 import itertools as itt
 import logging
 from abc import abstractmethod
-from typing import Callable, NamedTuple, Optional, Sequence, Tuple, Union
+from typing import Optional, Sequence, Tuple, Union
 
 import numpy
 import numpy as np
@@ -17,7 +17,7 @@ from .base import Model
 from ..triples import CoreTriplesFactory, TriplesFactory
 from ..triples.utils import tensor_to_df
 from ..typing import LabeledTriples, MappedTriples, ScorePack
-from ..utils import get_dropout_modules, is_cuda_oom_error
+from ..utils import is_cuda_oom_error
 
 __all__ = [
     "predict",
@@ -26,12 +26,6 @@ __all__ = [
     "get_head_prediction_df",
     "get_relation_prediction_df",
     "get_tail_prediction_df",
-    # Uncertainty prediction
-    "UncertainPrediction",
-    "predict_hrt_uncertain",
-    "predict_h_uncertain",
-    "predict_r_uncertain",
-    "predict_t_uncertain",
 ]
 
 logger = logging.getLogger(__name__)
@@ -673,7 +667,7 @@ def predict_triples_df(
     """
     Predict on labeled or mapped triples.
 
-    Example:
+    Example::
 
     >>> from pykeen.pipeline import pipeline
     >>> result = pipeline(dataset="nations", model="TransE")
@@ -734,248 +728,3 @@ def predict_triples_df(
         return tensor_to_df(tensor=triples, score=scores)
 
     return triples_factory.tensor_to_df(tensor=triples, score=scores)
-
-
-class UncertainPrediction(NamedTuple):
-    """A pair of predicted scores and corresponding uncertainty."""
-
-    #: The scores
-    score: torch.FloatTensor
-
-    #: The uncertainty, in the same shape as scores
-    uncertainty: torch.FloatTensor
-
-
-@torch.inference_mode()
-def _predict_uncertain(
-    model: Model,
-    batch: torch.LongTensor,
-    score_method: Callable[..., torch.FloatTensor],
-    num_samples: int,
-    slice_size: Optional[int] = None,
-) -> UncertainPrediction:
-    """
-    Predict with uncertainty estimates via Monte-Carlo dropout.
-
-    .. note::
-        the model has to contain at least one dropout layer.
-
-    .. note::
-        the model will be set to evaluation mode, and all dropout layers will be set to training mode
-
-    :param model:
-        the model used for predicting scores
-    :param batch:
-        the batch on which to predict. Its shape and content has to match what the `score_method` requires.
-    :param score_method:
-        the base score method to use (from `score_{hrt,h,r,t}`)
-    :param num_samples: > 1
-        The number of samples to use. More samples lead to better estimates, but increase memory requirements and
-        runtime.
-    :param slice_size: >0
-        The divisor for the scoring function when using slicing.
-
-    :return:
-        A tuple (score_mean, score_std) of the mean and std of the scores sampled from the dropout distribution.
-        The std may be interpreted as a measure of uncertainty.
-
-    :raises ValueError:
-        if the model does not contain dropout layers.
-    """
-    dropout_modules = get_dropout_modules(model)
-    if not dropout_modules:
-        raise ValueError(
-            "Model needs to contain at least one dropout layer to use the Monte-Carlo Dropout technique.",
-        )
-
-    # Enforce evaluation mode
-    model.eval()
-
-    # set dropout layers to training mode
-    for module in dropout_modules:
-        module.train()
-
-    kwargs = dict()
-    if slice_size is not None:
-        kwargs["slice_size"] = slice_size
-
-    # draw samples
-    batch = batch.to(model.device)
-    scores = torch.stack([score_method(batch, **kwargs) for _ in range(num_samples)], dim=0)
-    if model.predict_with_sigmoid:
-        scores = torch.sigmoid(scores)
-
-    # compute mean and std
-    return UncertainPrediction(score=scores.mean(dim=0), uncertainty=scores.std(dim=0))
-
-
-def predict_hrt_uncertain(
-    model: Model,
-    hrt_batch: torch.LongTensor,
-    num_samples: int = 5,
-) -> UncertainPrediction:
-    """
-    Calculate the scores with uncertainty quantification via Monto-Carlo dropout as proposed in [berrendorf2021]_.
-
-    .. seealso::
-        :func:`pykeen.models.Model.score_hrt`, :func:`_predict_uncertain`
-
-    .. note::
-        this method requires the model to have at least one dropout layer.
-
-    Example Usage:
-
-    .. code-block::
-
-        from pykeen.pipeline import pipeline
-        from pykeen.models.predict import predict_hrt_uncertain
-
-        result = pipeline(dataset="nations", model="ERMLPE")
-        prediction_with_uncertainty = predict_hrt_uncertain(
-            model=result.model,
-            hrt_batch=result.training.mapped_triples[0:8],
-        )
-
-    :param model:
-        the model used for predicting scores
-    :param hrt_batch: shape: (number of triples, 3), dtype: long
-        The indices of (head, relation, tail) triples.
-    :param num_samples: >1
-        the number of samples to draw
-
-    :return: shape: (number of triples, 1), dtype: float
-        The score for each triple, and an uncertainty score, where larger scores correspond to less certain
-        predictions.
-    """
-    return _predict_uncertain(
-        model=model,
-        batch=hrt_batch,
-        score_method=model.score_hrt,
-        num_samples=num_samples,
-    )
-
-
-def predict_h_uncertain(
-    model: Model,
-    rt_batch: torch.LongTensor,
-    num_samples: int = 5,
-    slice_size: Optional[int] = None,
-) -> UncertainPrediction:
-    """Forward pass using left side (head) prediction for obtaining scores of all possible heads.
-
-    This method calculates the score for all possible heads for each (relation, tail) pair, as well as an uncertainty
-    quantification.
-
-    Additionally, the model is set to evaluation mode.
-
-    .. note::
-
-        If the model has been trained with inverse relations, the task of predicting
-        the head entities becomes the task of predicting the tail entities of the
-        inverse triples, i.e., $f(*,r,t)$ is predicted by means of $f(t,r_{inv},*)$.
-
-    .. note::
-        this method requires the model to have at least one dropout layer.
-
-    :param model:
-        the model used for predicting scores
-    :param rt_batch: shape: (batch_size, 2), dtype: long
-        The indices of (relation, tail) pairs.
-    :param slice_size: >0
-        The divisor for the scoring function when using slicing.
-    :param num_samples: >1
-        the number of samples to draw
-
-    :return: shape: (batch_size, num_entities), dtype: float
-        For each r-t pair, the scores for all possible heads.
-    """
-    return _predict_uncertain(
-        model=model,
-        batch=rt_batch,
-        score_method=model.score_h_inverse if model.use_inverse_triples else model.score_h,
-        num_samples=num_samples,
-        slice_size=slice_size,
-    )
-
-
-def predict_r_uncertain(
-    model: Model,
-    ht_batch: torch.LongTensor,
-    num_samples: int = 5,
-    slice_size: Optional[int] = None,
-) -> UncertainPrediction:
-    """Forward pass using middle (relation) prediction for obtaining scores of all possible relations.
-
-    This method calculates the score for all possible relations for each (head, tail) pair, as well as an uncertainty
-    quantification.
-
-    Additionally, the model is set to evaluation mode.
-
-    .. note::
-        this method requires the model to have at least one dropout layer.
-
-    :param model:
-        the model used for predicting scores
-    :param ht_batch: shape: (batch_size, 2), dtype: long
-        The indices of (head, tail) pairs.
-    :param slice_size: >0
-        The divisor for the scoring function when using slicing.
-    :param num_samples: >1
-        the number of samples to draw
-
-    :return: shape: (batch_size, num_relations), dtype: float
-        For each h-t pair, the scores for all possible relations.
-    """
-    return _predict_uncertain(
-        model=model,
-        batch=ht_batch,
-        score_method=model.score_r,
-        num_samples=num_samples,
-        slice_size=slice_size,
-    )
-
-
-def predict_t_uncertain(
-    model: Model,
-    hr_batch: torch.LongTensor,
-    num_samples: int = 5,
-    slice_size: Optional[int] = None,
-) -> UncertainPrediction:
-    """Forward pass using right side (tail) prediction for obtaining scores of all possible tails.
-
-    This method calculates the score for all possible tails for each (head, relation) pair, as well as an uncertainty
-    quantification.
-
-    Additionally, the model is set to evaluation mode.
-
-    .. note::
-
-        We only expect the right side-side predictions, i.e., $(h,r,*)$ to change its
-        default behavior when the model has been trained with inverse relations
-        (mainly because of the behavior of the LCWA training approach). This is why
-        the :func:`predict_scores_all_heads` has different behavior depending on
-        if inverse triples were used in training, and why this function has the same
-        behavior regardless of the use of inverse triples.
-
-    .. note::
-        this method requires the model to have at least one dropout layer.
-
-    :param model:
-        the model used for predicting scores
-    :param hr_batch: shape: (batch_size, 2), dtype: long
-        The indices of (head, relation) pairs.
-    :param slice_size: >0
-        The divisor for the scoring function when using slicing.
-    :param num_samples: >1
-        the number of samples to draw
-
-    :return: shape: (batch_size, num_entities), dtype: float
-        For each h-r pair, the scores for all possible tails.
-    """
-    return _predict_uncertain(
-        model=model,
-        batch=hr_batch,
-        score_method=model.score_t,
-        num_samples=num_samples,
-        slice_size=slice_size,
-    )
