+++ tox.ini
@@ -109,6 +109,7 @@ commands =
     docstr-coverage --skip-private --skip-magic \
         src/pykeen/datasets/ \
         src/pykeen/models/predict.py \
+        src/pykeen/models/uncertainty.py \
         src/pykeen/utils.py \
         src/pykeen/datasets/ \
         src/pykeen/triples/utils.py \
@@ -125,6 +126,7 @@ deps =
 commands =
     flake8 \
         src/pykeen/models/predict.py \
+        src/pykeen/models/uncertainty.py \
         src/pykeen/models/nbase.py \
         src/pykeen/utils.py \
         src/pykeen/datasets/ \
